+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## unreleased
 
+- Add EC keys and parameters parsers
+
 ## v0.1.0
 
 *2015-11-27*